requirements.txt
anomalo==0.21.1
certifi==2024.2.2
charset-normalizer==3.3.2
dataclasses-json==0.5.14
deepmerge==1.1.1
fire==0.4.0
idna==3.6
iniconfig==2.0.0
Jinja2==3.1.3
MarkupSafe==2.1.5
marshmallow==3.21.1
mypy-extensions==1.0.0
packaging==24.0
pluggy==1.4.0
PyYAML==6.0.1
requests==2.31.0
setuptools==69.2.0
six==1.16.0
tenacity==8.2.3
termcolor==2.4.0
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
wheel==0.43.0
